move PackageUtilsTest tests to fitting location
