+++ java/code/src/com/redhat/rhn/domain/rhnpackage/test/PackageEvrTest.java
@@ -28,4 +28,77 @@ public class PackageEvrTest extends BaseTestCaseWithUser {
         assertEquals("1:2.3.4-5", evr1.toUniversalEvrString());
         assertEquals("1.2", evr2.toUniversalEvrString());
     }
+
+    /**
+     * Debian package versioning policy format: [epoch:]upstream_version[-debian_revision]
+     * Additional ':' and '-' characters are allowed in 'upstream_version'
+     * https://www.debian.org/doc/debian-policy/ch-controlfields.html#version
+     *
+     * Tests:
+     *   - 1:2.3~.4a+b-5
+     *   - 2.3~.4a+b-5
+     *   - 2.3~.4a+b
+     *   - 1:2.3~.4a+b
+     *   - 1:2.3~.4a+b-5+abc.6~
+     *   - 1:2-3-4-5
+     *
+     */
+    public void testParseDebianEvr() {
+
+        PackageEvr evr;
+
+        evr = PackageEvr.parseDebian("1:2.3~.4a+b-5");
+        assertEquals("1", evr.getEpoch());
+        assertEquals("2.3~.4a+b", evr.getVersion());
+        assertEquals("5", evr.getRelease());
+
+        evr = PackageEvr.parseDebian("2.3~.4a+b-5");
+        assertNull(evr.getEpoch());
+        assertEquals("2.3~.4a+b", evr.getVersion());
+        assertEquals("5", evr.getRelease());
+
+        evr = PackageEvr.parseDebian("2.3~.4a+b");
+        assertNull(evr.getEpoch());
+        assertEquals("2.3~.4a+b", evr.getVersion());
+        assertEquals("X", evr.getRelease());
+
+        evr = PackageEvr.parseDebian("1:2.3~.4a+b");
+        assertEquals("1", evr.getEpoch());
+        assertEquals("2.3~.4a+b", evr.getVersion());
+        assertEquals("X", evr.getRelease());
+
+        evr = PackageEvr.parseDebian("1:2.3~.4a+b-5+abc.6~");
+        assertEquals("1", evr.getEpoch());
+        assertEquals("2.3~.4a+b", evr.getVersion());
+        assertEquals("5+abc.6~", evr.getRelease());
+
+        evr = PackageEvr.parseDebian("2-3-4-5");
+        assertNull(evr.getEpoch());
+        assertEquals("2-3-4", evr.getVersion());
+        assertEquals("5", evr.getRelease());
+    }
+
+    public void testParseRpmEvr() {
+        PackageEvr evr;
+
+        evr = PackageEvr.parseRpm("1:1.2.3-4.5");
+        assertEquals("1", evr.getEpoch());
+        assertEquals("1.2.3", evr.getVersion());
+        assertEquals("4.5", evr.getRelease());
+
+        evr = PackageEvr.parseRpm("1.2.3-4.5");
+        assertNull(evr.getEpoch());
+        assertEquals("1.2.3", evr.getVersion());
+        assertEquals("4.5", evr.getRelease());
+
+        evr = PackageEvr.parseRpm("1:1.2.3-4.5");
+        assertEquals("1", evr.getEpoch());
+        assertEquals("1.2.3", evr.getVersion());
+        assertEquals("4.5", evr.getRelease());
+
+        evr = PackageEvr.parseRpm("1.2.3");
+        assertNull(evr.getEpoch());
+        assertEquals("1.2.3", evr.getVersion());
+        assertEquals("", evr.getRelease());
+    }
 }
